fix build error on macos

+++ Framework/Common/GraphicsManager.hpp
@@ -36,22 +36,22 @@ class GraphicsManager : _implements_ IRuntimeModule {
 
     virtual void DrawBatch(const Frame& frame) {}
 
-    virtual int32_t GenerateCubeShadowMapArray(const uint32_t width,
+    virtual intptr_t GenerateCubeShadowMapArray(const uint32_t width,
                                                const uint32_t height,
                                                const uint32_t count) {
         return 0;
     }
-    virtual int32_t GenerateShadowMapArray(const uint32_t width,
+    virtual intptr_t GenerateShadowMapArray(const uint32_t width,
                                            const uint32_t height,
                                            const uint32_t count) {
         return 0;
     }
     virtual void BeginShadowMap(const int32_t light_index,
-                                const int32_t shadowmap, const uint32_t width,
+                                const intptr_t shadowmap, const uint32_t width,
                                 const uint32_t height,
                                 const int32_t layer_index, const Frame& frame) {
     }
-    virtual void EndShadowMap(const int32_t shadowmap,
+    virtual void EndShadowMap(const intptr_t shadowmap,
                               const int32_t layer_index) {}
     virtual void SetShadowMaps(const Frame& frame) {}
 
@@ -96,17 +96,17 @@ class GraphicsManager : _implements_ IRuntimeModule {
                               const Matrix4X4f& trans, const Vector3f& color) {}
     virtual void DrawTriangleStrip(const PointList& vertices,
                                    const Vector3f& color) {}
-    virtual void DrawTextureOverlay(const int32_t texture, const float vp_left,
+    virtual void DrawTextureOverlay(const intptr_t texture, const float vp_left,
                                     const float vp_top, const float vp_width,
                                     const float vp_height) {}
     virtual void DrawTextureArrayOverlay(
-        const int32_t texture, const float layer_index, const float vp_left,
+        const intptr_t texture, const float layer_index, const float vp_left,
         const float vp_top, const float vp_width, const float vp_height) {}
-    virtual void DrawCubeMapOverlay(const int32_t cubemap, const float vp_left,
+    virtual void DrawCubeMapOverlay(const intptr_t cubemap, const float vp_left,
                                     const float vp_top, const float vp_width,
                                     const float vp_height, const float level) {}
     virtual void DrawCubeMapArrayOverlay(
-        const int32_t cubemap, const float layer_index, const float vp_left,
+        const intptr_t cubemap, const float layer_index, const float vp_left,
         const float vp_top, const float vp_width, const float vp_height,
         const float level) {}
     virtual void ClearDebugBuffers() {}
